Enemy Distribution
Added another enemy based on enemy 4 that is smaller, moves much faster, has a light trail and always drops a powerup. Enemy 5 only has a 1% chance of spawning to make up for this, and all the other enemies have had their spawn chance adjusted. (exact percentages are on Miro Board)